Changes based on PR suggestions

+++ byte_fifo.c
@@ -41,13 +41,13 @@ int16_t byte_fifo_is_full(const struct byte_fifo_t* const fifo)
     return (fifo->n_elements == fifo->size);
 }
 
-int16_t byte_fifo_write(struct byte_fifo_t* const fifo, const uint8_t* const src, uint16_t len)
+int32_t byte_fifo_write(struct byte_fifo_t* const fifo, const uint8_t* const src, uint16_t len)
 {
     RETURN_IF(NULL == fifo, BYTE_FIFO_NULLPTR);
     RETURN_IF(NULL == fifo->data, BYTE_FIFO_NULLPTR);
     RETURN_IF(NULL == src, BYTE_FIFO_NULLPTR);
 
-    int16_t n_bytes_written = 0;
+    int32_t n_bytes_written = 0;
     while (len > 0)
     {
         if (fifo->n_elements < fifo->size)
@@ -71,14 +71,14 @@ int16_t byte_fifo_write(struct byte_fifo_t* const fifo, const uint8_t* const src
     return n_bytes_written;
 }
 
-int16_t byte_fifo_overwrite(struct byte_fifo_t* const fifo, const uint8_t* const src, uint16_t len)
+int32_t byte_fifo_overwrite(struct byte_fifo_t* const fifo, const uint8_t* const src, uint16_t len)
 {
     RETURN_IF(NULL == fifo, BYTE_FIFO_NULLPTR);
     RETURN_IF(NULL == fifo->data, BYTE_FIFO_NULLPTR);
     RETURN_IF(NULL == src, BYTE_FIFO_NULLPTR);
 
-    int16_t n_bytes_written = 0;
-    int16_t n_bytes_overwritten = 0;
+    int32_t n_bytes_written = 0;
+    int32_t n_bytes_overwritten = 0;
     while (len > 0)
     {
         uint16_t write_index = fifo->write_index;
@@ -102,14 +102,14 @@ int16_t byte_fifo_overwrite(struct byte_fifo_t* const fifo, const uint8_t* const
     return n_bytes_overwritten;
 }
 
-int16_t byte_fifo_read(struct byte_fifo_t* const fifo, uint8_t* const dest, int16_t len)
+int32_t byte_fifo_read(struct byte_fifo_t* const fifo, uint8_t* const dest, int16_t len)
 {
     RETURN_IF(NULL == fifo, BYTE_FIFO_NULLPTR);
     RETURN_IF(NULL == fifo->data, BYTE_FIFO_NULLPTR);
     RETURN_IF(NULL == dest, BYTE_FIFO_NULLPTR);
     RETURN_IF(len < 0, BYTE_FIFO_INVALID_PARAM);
 
-    int16_t n_bytes_read = 0;
+    int32_t n_bytes_read = 0;
     while (n_bytes_read < len)
     {
         if (fifo->n_elements > 0)
